% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_BS.R
\name{CNORprob_BS}
\alias{CNORprob_BS}
\title{CNORprob_BS}
\usage{
CNORprob_BS(model, CNOlist, estim, res, BS_Type, BS_Round, AddSD = 0.05,
  SeedNr = NULL)
}
\description{
Perform Bootstrapping with random
}
