% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_edgeKO.R
\name{CNORprob_edgeKO}
\alias{CNORprob_edgeKO}
\title{CNORprob_edgeKO}
\usage{
CNORprob_edgeKO(model, CNOlist, estim, res)
}
\description{
Systematically knock-out each single edge from the network, run optimisation of each network variants and compare models with fitting costs and AIC
}
