% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_mapModel.R
\name{CNORprob_mapModel}
\alias{CNORprob_mapModel}
\title{CNORprob_mapModel}
\usage{
CNORprob_mapModel(model, CNOlist, estim, res)
}
\description{
Map optimised parameters from CNORprob back to each reaction in model$reacID to be compatible with model plotting function in the CellNOptR package
}
