% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_nodeKO.R
\name{CNORprob_nodeKO}
\alias{CNORprob_nodeKO}
\title{CNORprob_nodeKO}
\usage{
CNORprob_nodeKO(model, CNOlist, estim, res)
}
\description{
Systematically knock-out each single node in the network, run optimisation of each network variant and compare models with fitting costs and AIC
}
