% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_optimise.R
\name{CNORprob_optimise}
\alias{CNORprob_optimise}
\title{CNORprob_optimise}
\usage{
CNORprob_optimise(estim, optRound = 3, SaveOptResults = TRUE,
  DispOptResults = TRUE)
}
\description{
An optimisation interface between model descriptions 'estim' and the optimiser 'rsolnp'
}
