% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_plotFit.R
\name{CNORprob_plotFit}
\alias{CNORprob_plotFit}
\title{CNORprob_plotFit}
\usage{
CNORprob_plotFit(model, CNOlist, estim, res, show = TRUE, plotPDF = TRUE,
  tag = NULL, plotParams = list(cex = 0.8, cmap_scale = 1))
}
\description{
Plot optimisation results against the experimental data
}
