% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_simulate.R
\name{CNORprob_simulate}
\alias{CNORprob_simulate}
\title{CNORprob_simulate}
\usage{
CNORprob_simulate(estim, SimParams, SimIterations)
}
\description{
Simulation script with a defined set of parameters 'initial guess or optimised'
}
