% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_writeConstraint.R
\name{CNORprob_writeConstraint}
\alias{CNORprob_writeConstraint}
\title{CNORprob_writeConstraint}
\usage{
CNORprob_writeConstraint(Interactions, HLbound, state_names,
  HardConstraint, LPSA = NULL)
}
\description{
Write constraints for probabilistic logic models before passing to the optimiser
}
