% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_Prob.R
\name{preprocessing_Prob}
\alias{preprocessing_Prob}
\title{preprocessing_Prob}
\usage{
preprocessing_Prob(data = NULL, model, cutNONC = TRUE,
  compression = TRUE, expansion = FALSE, ignoreList = NA,
  maxInputsPerGate = 2, verbose = TRUE)
}
\description{
preprocessing_Prob
}
