% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCNORprob.R
\name{runCNORprob}
\alias{runCNORprob}
\title{runCNORprob}
\usage{
runCNORprob(model = NULL, data = NULL, CNORprob_Example = 1,
  ProbCompression = F, ProbCutNONC = F, ProbExpandOR = F,
  ORlist = NULL, ProbHardConstraint = F, ProbForce = F,
  optRound_optim = 3, L1Reg = 1e-04, MaxTime = 180, HLbound = 0.5,
  SSthresh = 2e-16, printCost = 0, PlotIterations = 1,
  SaveOptResults = TRUE, rho = 1, outer.iter = 100,
  inner.iter = 100, delta = 1e-07, tol = 1e-08, trace = 1,
  Analysis_edgeKO = T, Analysis_nodeKO = T, Analysis_LPSA = T,
  Analysis_BS = T, optRound_analysis = 1, LPSA_Increments = 2,
  BS_Type = 1, BS_Round = 5)
}
\arguments{
\item{model}{A filename of prior knowledge network (PKN) in the SIF format}

\item{data}{A measurement filename in the MIDAS format}

\item{CNORprob_Example}{A variable to select CNORprob examples (1-4)}

\item{ProbCompression}{A variable to define whether to perform a compression step}

\item{ProbCutNONC}{A variable to define whether to cut non-observable non-controllable node from PKN}

\item{ProbExpandOR}{A variable to define whether to expand OR gate in the CNOR format}

\item{ORlist}{A list of interaction with OR gate (leave NULL is none)}

\item{ProbHardConstraint}{Apply hard constraints on probability parameters (sum of all activatory edges = 1)}

\item{ProbForce}{A variable to define whether to force all single activatory edge to a node to always be 1}

\item{optRound_optim}{Number of rounds for optimisation}

\item{L1Reg}{Weight for L1-regularisation}

\item{MaxTime}{Time for each round of optimisation (seconds)}

\item{HLbound}{A cut-off for high and low weights}

\item{SSthresh}{A cut-off for states' difference to define whether steady-state is reached}

\item{printCost}{A variable to select whether to print or not to print intermediate fitting cost (0,1)}

\item{PlotIterations}{Number of rounds for optimisation to generate plots}

\item{SaveOptResults}{A variable to define whether to generate reports from optimisation (TRUE,FALSE)}

\item{rho}{Penalty weighting scaler / default = 1}

\item{outer.iter}{Maximum major iterations / default = 400}

\item{inner.iter}{Maximum minor iterations / default = 800}

\item{delta}{Relative step size eval. / default = 1e-7}

\item{tol}{Relative tol. for optim. / default = 1e-8}

\item{Analysis_edgeKO}{Perform edge KO analysis (T/F)}

\item{Analysis_nodeKO}{Perform node KO analysis (T/F)}

\item{Analysis_LPSA}{Perform local parameter sensitivity analysis (T/F)}

\item{Analysis_BS}{Perform bootstrapping analysis (T/F)}

\item{optRound_analysis}{Number of rounds for optmisation in each post-optimisation analysis}

\item{LPSA_Increments}{Number of increments in LPSA analysis}

\item{BS_Type}{Type of Bootstrapping (1=resample with replacement from residual; 2=resampling from mean & variant)}

\item{BS_Round}{Number of rounds for bootstrapping}

\item{iter}{Print objfunc every itereration / default = 1}
}
\description{
A one-linear wrapper of the CNORprob pipeline
}
