% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNORprob_simulate.R
\name{CNORprob_simulate}
\alias{CNORprob_simulate}
\title{CNORprob_simulate}
\usage{
CNORprob_simulate(estim, SimParams, SimIterations)
}
\description{
Simulate CNORprob model with a defined set of parameters: either 'initial guess' or 'optimised'
}
